+++ examples/example/App.tsx
@@ -1,0 +1,96 @@
+import React from 'react'
+import {
+	View,
+	Text as DripText,
+	createThemedComponent,
+	useResponsiveValue,
+	DripsyProvider,
+	setDripsyOptions,
+	Container,
+} from 'dripsy'
+import { Text } from 'react-native'
+
+setDripsyOptions({ ssr: true })
+
+const theme = {
+	useBodyStyles: false,
+	useLocalStorage: false,
+	useCustomProperties: false,
+	useColorSchemeMediaQuery: false,
+	colors: {
+		primary: '#41b87a',
+		secondary: 'black',
+		background: 'white',
+		red: 'red',
+		green: 'green',
+		blue: 'blue',
+	},
+	text: {
+		primary: {
+			fontSize: 40,
+		},
+	},
+}
+
+const G = createThemedComponent(Text, {
+	themeKey: 'text',
+})
+
+const ResponsiveSquare = () => {
+	// Return literal values:
+	const textColor = useResponsiveValue(['red', 'green', 'blue'])
+	// Or provide a function to access theme values:
+	const squareColor = useResponsiveValue((theme) => [
+		theme?.colors?.blue,
+		theme?.colors?.red,
+		theme?.colors?.green,
+	])
+
+	return (
+		<View
+			sx={{
+				width: [100, 120, 150],
+				height: [100, 120, 150],
+				bg: squareColor,
+				mt: 1,
+			}}
+		>
+			<DripText sx={{ color: textColor }}>Hello</DripText>
+		</View>
+	)
+}
+
+export default function App() {
+	return (
+		<DripsyProvider theme={theme} options={{ ssr: true }}>
+			<Container>
+				<View
+					sx={{
+						backgroundColor: () => ['primary', 'white'],
+						height: [400, 800],
+					}}
+				>
+					<DripText>joi</DripText>
+					<G variant="primary">Hey</G>
+					<G>Hi!</G>
+					<View
+						sx={{ bg: 'white' }}
+						style={{
+							shadowColor: '#000',
+							shadowOffset: {
+								width: 0,
+								height: 2,
+							},
+							shadowOpacity: 0.25,
+							shadowRadius: 3.84,
+							// elevation: 5,
+						}}
+					>
+						<Text>Card</Text>
+					</View>
+					<ResponsiveSquare />
+				</View>
+			</Container>
+		</DripsyProvider>
+	)
+}
