+++ examples/example/metro.config.js
@@ -1,0 +1,4 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const { createMetroConfiguration } = require('expo-yarn-workspaces')
+
+module.exports = createMetroConfiguration(__dirname)
