+++ examples/example/web-build/register-service-worker.js
@@ -1,13 +1,13 @@
 /* eslint-env browser */
 
 if ('serviceWorker' in navigator) {
-  window.addEventListener('load', function () {
+  window.addEventListener('load', function() {
     navigator.serviceWorker
       .register('/expo-service-worker.js', { scope: '/' })
-      .then(function (info) {
+      .then(function(info) {
         // console.info('Registered service-worker', info);
       })
-      .catch(function (error) {
+      .catch(function(error) {
         console.info('Failed to register service-worker', error);
       });
   });
