+++ examples/next-example/babel.config.js
