+++ examples/next-example/next-env.d.ts
