+++ lerna.json
@@ -1,0 +1,20 @@
+{
+	"packages": ["packages/*", "example"],
+	"npmClient": "yarn",
+	"useWorkspaces": true,
+	"version": "independent",
+	"command": {
+		"publish": {
+			"allowBranch": "master",
+			"conventionalCommits": true,
+			"createRelease": "github",
+			"message": "Publish",
+			"ignoreChanges": [
+				"**/__fixtures__/**",
+				"**/__tests__/**",
+				"**/*.md",
+				"**/example/**"
+			]
+		}
+	}
+}
