--- next-example/package.json
@@ -1,25 +1,0 @@
-{
-  "dependencies": {
-    "dripsy": "../dripsy-v0.6.0-fresnel-2.1.tgz",
-    "expo": "37.0.7",
-    "next": "9.4.4",
-    "next-compose-plugins": "^2.2.0",
-    "next-transpile-modules": "3.3.0",
-    "react": "16.9.0",
-    "react-dom": "16.9.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
-    "react-native-web": "0.11.7"
-  },
-  "devDependencies": {
-    "@babel/core": "7.9.0",
-    "@expo/next-adapter": "2.1.0"
-  },
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo web",
-    "eject": "expo eject"
-  },
-  "private": true
-}
