+++ packages/dripsy/package.json
@@ -1,0 +1,104 @@
+{
+	"name": "dripsy",
+	"version": "1.2.1",
+	"description": "🍷 A super-simple responsive design system for React Native Web.",
+	"main": "lib/commonjs/index.js",
+	"module": "lib/module/index.js",
+	"types": "lib/typescript/src/index.d.ts",
+	"react-native": "src/index.tsx",
+	"source": "src/index.tsx",
+	"files": [
+		"src",
+		"lib"
+	],
+	"scripts": {
+		"test": "jest",
+		"typescript": "tsc --noEmit",
+		"lint": "eslint --ext .js,.ts,.tsx .",
+		"prepare": "bob build",
+		"release": "release-it",
+		"example": "yarn --cwd example",
+		"pods": "node -e \"process.exit(require('os').platform() !== 'darwin')\" && test -d example/ios && cd example/ios && pod install || printf ''",
+		"bootstrap": "yarn example && yarn && yarn pods",
+		"local": "yarn prepare && yarn pack && cd example && yarn add ../dripsy-v0.6.0-fresnel-2.1.tgz && yarn start -c",
+		"ssr": "yarn prepare && yarn pack && cd next-example && yarn add ../dripsy-v0.6.0-fresnel-2.1.tgz && yarn next"
+	},
+	"keywords": [
+		"react-native",
+		"ios",
+		"android"
+	],
+	"repository": "https://github.com/nandorojo/expo-theme-ui",
+	"author": "Fernando Rojo <frojo@sas.upenn.edu> (https://github.com/nandorojo)",
+	"license": "MIT",
+	"bugs": {
+		"url": "https://github.com/nandorojo/expo-theme-ui/issues"
+	},
+	"homepage": "https://github.com/nandorojo/expo-theme-ui#readme",
+	"devDependencies": {
+		"@commitlint/config-conventional": "^8.3.4",
+		"@react-native-community/bob": "^0.11.2",
+		"@react-native-community/eslint-config": "^1.1.0",
+		"@release-it/conventional-changelog": "^1.1.4",
+		"@types/jest": "^25.2.1",
+		"@types/react": "^16.9.43",
+		"@types/react-native": "^0.63.2",
+		"commitlint": "^8.3.5",
+		"eslint-config-nando": "^1.0.9",
+		"jest": "^26.0.1",
+		"metro-react-native-babel-preset": "^0.61.0",
+		"release-it": "^13.5.8",
+		"typescript": "^3.9.7"
+	},
+	"jest": {
+		"preset": "react-native",
+		"modulePathIgnorePatterns": [
+			"<rootDir>/example/node_modules",
+			"<rootDir>/lib/"
+		]
+	},
+	"eslintIgnore": [
+		"node_modules/",
+		"lib/"
+	],
+	"release-it": {
+		"git": {
+			"commitMessage": "chore: release ${version}",
+			"tagName": "v${version}"
+		},
+		"npm": {
+			"publish": true
+		},
+		"github": {
+			"release": true
+		},
+		"plugins": {
+			"@release-it/conventional-changelog": {
+				"preset": "angular"
+			}
+		}
+	},
+	"@react-native-community/bob": {
+		"source": "src",
+		"output": "lib",
+		"targets": [
+			"commonjs",
+			"module",
+			"typescript"
+		]
+	},
+	"peerDependencies": {
+		"react": "16.9.0",
+		"react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz"
+	},
+	"dependencies": {
+		"@artsy/fresnel": "^1.2.0",
+		"@expo/html-elements": "^0.0.0-alpha.5",
+		"@react-native-community/hooks": "^2.6.0",
+		"@theme-ui/core": "^0.4.0-rc.1",
+		"@theme-ui/css": "^0.4.0-rc.1",
+		"hoist-non-react-statics": "^3.3.2",
+		"theme-ui": "^0.4.0-rc.3"
+	},
+	"sideEffects": false
+}
