+++ packages/dripsy/src/index.tsx
@@ -1,4 +1,3 @@
-
 export * from './components'
 export * from './css/create-themed-component'
 export * from './provider'
