+++ packages/dripsy/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../tsconfig",
+	"compilerOptions": {
+		"paths": {},
+		"outDir": "./lib/typescript"
+	}
+}
